requirements.txt
Django
djangorestframework
python-dotenv
drf-spectacular
djangorestframework-simplejwt
